requirements.txt
django
django-cors-headers
djangorestframework==3.11.0
gunicorn
whitenoise
django-heroku
boto3
Pillow
django-allauth
django-rest-auth